moved launch era class to be in data package
