+++ app/src/main/java/com/spacex_mvvm/features/launchlist/usecase/ViewLaunchesUseCase.kt
@@ -3,7 +3,9 @@ package com.spacex_mvvm.features.launchlist.usecase
 import com.spacex_mvvm.data.Resource
 import com.spacex_mvvm.data.mapData
 import com.spacex_mvvm.data.repositories.launches.LaunchesRepository
-import com.spacex_mvvm.features.launchlist.view.LaunchEra
+import com.spacex_mvvm.data.repositories.launches.model.LaunchEra
+import com.spacex_mvvm.features.launchlist.model.LaunchListItem
+import com.spacex_mvvm.features.launchlist.model.LaunchListItemsMapper
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
